WebDAV: find default calendar
This patch implements a dynamic search for the "default" calendar
of a user, starting with the user's principal URL:
https://caldav.calendar.yahoo.com/dav/<user>/
https://www.google.com/calendar/dav/<user>/user/

Stops when:
 - current path is calendar collection (= contains VEVENTs)
Gives up:
 - when running in circles
 - nothing else to try out
 - tried 10 times
Follows:
 - CalDAV calendar-home-set
 - collections

This implies that a specific calendar can be selected by using its URL
directly. The search for "default" calendar is a bit random, because
it depends on the naming of calendars on the server: calendars with
lexicographically lower name are tested (and thus found) first.