CalDAV: fixes for inserting items into Yahoo
Content-Type must be set, otherwise Yahoo CalDAV rejects the
items with "Bad Request".

The resource name must match the UID + .ics. It is okay to
escape special characters like @ in the PUT, even though the
server doesn't do the same in PROPFIND.

For change tracking we need a canonical form for DAV luids. Always
unescaping the relative path component is safe. For absolute paths, we
use the value as suggested by the server. There's no risk of mismatch
here, because we never create such a luid ourself.