slow sync strategy: allow client and server updates for SyncEvolution clients
Merging between SyncEvolution instances is expected to work
well, so allow the engine to update items on both
sides also in the "non-first" slow sync case. Normally it
would only pair items, but not touch data on either side.

This is turned off by default for other clients
because small changes would propagate to other
clients of the server. But it also prevents getting truly
different items into sync during a slow sync.

Hopefully with SyncEvolution as client we'll have not much such
changes, so it makes sense to enable the updating.