local sync: rely on iCalendar 2.0 UID/RECURRENCE-ID for pairing
If both sides support iCalendar 2.0, then UID/RECURRENCE-ID should be
used for pairing instead of looking at data. The other user-visible
fields are still relevant and should be merged and updated. The new
Synthesis compare="scripted" achieves the later.

This patch enables this mode for local sync, which is assumed to use
iCalendar 2.0. This assumption and the way how the different compare
values are implemented (global search/replace via a poor man's regex)
is a hack: it would be much nicer to have this mode configurable for
each peer and source or even recognize it at runtime.

But runtime detection is unlikely to work without fundamental changes
in libsynthesis because the field list must be configured before the
peer is even known. One solution might be to evaluate the compare level
later at runtime, instead of at config parsing time.