calendar fields: ignore changes to alarm UID
The alarm UID (= X-EVOLUTION-ALARM-UID) is not user visible, so
changes in it can be ignored, aka not relevant = compare="never".

During conflict resolution it is better to not remove an existing
value: merge="fillempty". A better solution would be to have accurate
information whether a peer stores it (not done correctly in some
local sync scenarios).