client-test: workaround for test failure in Client::Source::ical20::testLinkedItemsRemoveNormal (BMC #10906)
Adding content verification to client-test
Client::Sync::ical20::testLinkedItemsRemoveNormal revealed a bug: when removing
the child event ("detached recurrence"), the expected outcome to have just the
original parent, as if the child had never existed.

Instead Evolution Data Server adds an "EXDATE" entry for the day when the child
recurred. This is a side effect of e_cal_remove_object_with_mod().

I consider the cure for this problem (replace e_cal_remove_object_with_mod()
with something like "remove all events, re-add only the relevant ones") worse
than the original problem:
- performance (several EDS calls instead of just one)
- code complexity
- common case (remove all events in a series) becomes slow and
  complex, whereas the unusual case (remove child) which needs
  the fix is very unlikely

To solve the nightly test failures, I preferred  to hack the test so that it adds
the EXDATE to the expected data in this particular case (easier than hacking
synccompare and more specific).