LogRedirect: allow user to remove error messages from stdout via SYNCEVOLUTION_SUPPRESS_ERRORS
The logic of printing all text containing the word "error" to stdout
can clutter up the output with know errors or simply false
positives (text which is not a real error).

Setting SYNCEVOLUTION_SUPPRESS_ERRORS to a newline-separated list
of lines will hide all text containing one of these lines. The text
is still logged.