client-test: get more tests to work in server mode
CLIENT_TEST_MODE=server enabled some tweaks for testing syncing with a
phone, but it did not go far enough. Tests like
Client::Sync::testRefreshFromServerSemantic were written under the
assumption that the peer is a SyncML server, and thus either used the
wrong refresh direction or expected the wrong result.

Instead of renaming all tests and the "client-test" program itself,
this patch merely changes the data direction in the affected tests
when running with CLIENT_TEST_MODE=server. In that mode, the meaning
of "server" and "client" are thus reverted. Potentially confusing :-/