XML config: avoid duplicating SUMMARY as DESCRIPTION
When sending a calendar event or task to a peer, the DESCRIPTION is now
only set to the SUMMARY in cases where the peer is known to not support
the SUMMARY property. This is an improvement over the original logic
which always duplicated the text.

When receiving such an item, a DESCRIPTION which is identical to the
SUMMARY is removed. This restores the original logic from upstream
Synthesis configs. Since commit
ca22c2d4c45136d543419bfb98f356aef1e3ff88, SyncEvolution avoided empty
DESCRIPTION fields by setting the DESCRIPTION to the SUMMARY because
libical was unhappy about it; but that was already successfully
avoided by another change in that commit, so it makes sense to avoid
storing duplicate data, should the peer send it (from a previous sync
or because it always sends the same text in both properties).