autotools: fixed some issues with building release version
First, the "grep" command which was meant to detect tags older
than HEAD did not work. Did not matter in practice because a full
version including date and hash was still produced.

Second, when multiple tags were assigned to the same hash (as in a
Synthesis library which doesn't change between prerelease and
release), the tag picked was random. In the SyncEvolution 1.1.1
release it happened to pick 1.1.0.99.1, requiring manual workarounds
to get a clean version.

The fix is to pick the alphabetically latest tag, which (with our
naming scheme) typically is the highest one.