syncevo-dbus-server: expect error when closing connection after failure
Instead of logging a
  org.freedesktop.DBus.Error.UnknownMethod: Method "Close" with signature "bs" on interface "org.syncevolution.Connection" doesn't exist
error, catch this error and only print it at debug level.

This is caused by syncevo-dbus-server removing the connection
right away after sending the Abort signal, so when Close() is
called on it, the instance no longer exists. The D-Bus error
is a bit misleading.

Not calling Close() would be the alternative, but determining
in advance when Close() is not necessary is hard, so better try in
all cases.