syncevo-dbus-server: adapted to ConnMan D-Bus API change (BMC #11998)
Between MeeGo 1.1 and 1.2, the ConnMan D-Bus API has changed from
org.moblin to net.connman as identifier for the service.

This patch adapts to that change with a string replace. No attempt
is made to stay compatible with previous ConnMan releases because
it is expected that the main target, MeeGo 1.2, will release with
updated SyncEvolution and ConnMan versions.