config classes: added reload()
Reloading properties of a node or a complete tree is necessary
when some other operation modified the background storage. One
use cases is rewriting .ini files during a migration while the
a SyncConfig has an active config tree for that config.

This is not a general solution to the problem of detecting
changes to the background store. Reloading must be triggered by
the code which made the change there.