nightly testing: MemoToo uses local time for recurring events
MemoToo converts UTC time into the local time configured for a user
when the event recurrs. That transformation is flawed, but let's
ignore it in the nightly testing by using test data which was
transformed the same way.

Note that the test account must be configured for the German time
zone, otherwise test data and MemoToo still won't match.