client-test: avoid special characters in copy/update tests
Previously test data for insert/copy/update tests was the same
as the template for creating items. But it was used without replacing
<<REVISION>> or REVISION, which confused some servers (for example,
Memotoo and memos).

This patch cleans this up such that the normal test data is simpler
and only the template items have a revision marker. Also changed
the templates so that the revision marker is <<REVISION>> across
the board (including memos) and can occur more than once, which is
used to get unique contact names.