Yahoo! Contacts: not stable enough to be enabled
Non-ASCII characters and pictures get lost. X- Extensions can be
uploaded and are shown on the server, but are not returned by it.
These issues are all on the server side and it is unclear how to
work around them.

Therefore this patch disables contact support in the Yahoo! profile.
Can be enabled manually for testing purposes.