client-test: added Client::Sync::*::testTimeout
This test verifies that timeout detection works for SyncML/HTTP and
local sync. The test does not depend on the source which is active,
so to cut down redundant testing it registers itself only for
the first source in CLIENT_TEST_SOURCES.

The implementation of the test creates a dead listening socket which
the HTTP transports can connect to without ever getting a reply. Then
the syncURL for outgoing HTTP is redirected towards that dead port.
The expectation is that the overall sync times out roughly after 10
seconds.